requirements.txt
Babel==2.3.4
Django==1.10.6
Faker==0.7.7
Pillow==4.0.0
PyJWT==1.4.2
Pygments==2.2.0
URLObject==2.4.2
Unidecode==0.04.20
amqp==2.1.4
awesome-slugify==1.6.5
beautifulsoup4==4.5.3
billiard==3.5.0.2
celery==4.0.2
decorator==4.0.11
defusedxml==0.5.0
django-appconf==1.0.2
django-autofixture==0.12.1
django-autoslug==1.9.3
django-bootstrap3==8.1.0
django-braces==1.11.0
django-celery-email==1.1.5
django-ckeditor==5.2.1
django-compressor==2.1.1
django-countries==4.0
django-crispy-forms==1.6.1
django-debug-panel==0.8.3
django-debug-toolbar==1.6
django-extensions==1.7.6
django-grappelli==2.9.1
django-htmlmin==0.10.0
django-imagekit==3.3
django-mathfilters==0.4.0
django-model-utils==2.6.1
django-recaptcha==1.2.1
django-redis==4.7.0
django-watson==1.3.0
django-widget-tweaks==1.4.1
djangorestframework==3.5.3
flower==0.9.1
gunicorn==19.6.0
html5lib==0.999999999
ipython==5.2.2
ipython-genutils==0.1.0
kombu==4.0.2
oauthlib==2.0.1
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
pilkit==2.0
prompt-toolkit==1.0.13
psycopg2==2.6.2
ptyprocess==0.5.1
pycountry==17.01.08
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2016.10
rcssmin==1.0.6
redis==2.10.5
regex==2017.02.08
requests==2.13.0
requests-oauthlib==0.8.0
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
sqlparse==0.2.3
tornado==4.2
traitlets==4.3.2
vine==1.1.3
wcwidth==0.1.7
webencodings==0.5
